Feat: 최근검색어 클릭 시 결과 페이지 이동 #57
